requirements.txt
Django==3.2.16
<<<<<<< HEAD
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
=======
pep8-naming==0.13.3
pytest==7.1.3
pytest-django==4.5.2
pytest-pythonpath==0.7.3
flake8==5.0.4
flake8-docstrings==1.7.0
>>>>>>> origin/main
